% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{wilcoxon.compute}
\alias{wilcoxon.compute}
\title{Wilcoxon Test and Plotting Function}
\usage{
wilcoxon.compute(data, measure.name, folder.path)
}
\arguments{
\item{data}{A data frame where the first column is an identifier and the following 
columns represent the two groups to be compared. The two columns for comparison should be 
in the second and third positions.}

\item{measure.name}{A string representing the name of the measure, used for naming 
output files.}

\item{folder.path}{A string representing the path to the folder where output files 
will be saved.}
}
\value{
A data frame containing the following columns:
\item{zeros}{Number of zero differences between the two columns.}
\item{statistic}{The test statistic from the Wilcoxon test.}
\item{parameter}{The parameter used in the Wilcoxon test.}
\item{p.value}{The p-value from the Wilcoxon test.}
\item{location.shift}{The location shift parameter used in the Wilcoxon test.}
\item{alternative}{The alternative hypothesis used in the Wilcoxon test.}
\item{method}{The method used for the Wilcoxon test.}
}
\description{
This function performs a paired Wilcoxon test on two columns of a data frame, 
saves the results to a text file, and generates plots to visualize the results.

The Wilcoxon test is a non-parametric statistical test used to compare two paired samples. 
This function computes the test and provides a summary of the results, including 
the number of zero differences, the test statistic, and the p-value. 
It also generates plots to help visualize the distribution and results of the test.
}
\examples{
# Example usage:
# Suppose 'my_data' is a data frame with columns 'Identifier', 'Group1', and 'Group2'.
# The first column should be named 'Identifier' and the columns to be compared should be 
# the second and third columns.
result <- wilcoxon.compute(data = my_data, measure.name = "accuracy", folder.path = "/results")

}
